Ai and Player Character work
Implemented the drop feature:

When enemies die, they will drop their own items. The items will have gravity, but will not react to the player character or the other enemy AIs.
These items can be picked up by the player, and used like any other item in the world.

Also added a feature for the player character. If they select the item, and press the drop button, the same physics will be applied to them. Dropping an item will remove it from the inventory and spawn it into the world, ahead of the player.

Also added an early version of the "Randomize Gear" and "Randomize Items" function through an enum. The only currently working way is to give each actor its own set of items by hand, but in the future, i wish to have an option for random gear and item spawning on each actor.

Also gave a specific object response to interactables. This way they have their own type, this makes it easier to specify the physics options only to the items