Ai and player character work
Implemented a way for the ai to have random attack animations through the enemy type enum. The way it works:

- the first time the ai (sword and shield for now) goes in chase mode, it will do either 2 normal attacks when reaching the player, or 2 jump attacks. These 2 jump attacks will not be repeated again.

- after this, the ai will create 2 arrays of attack types (inward hit and outward hit). These two attack types will interchange each times it attacks. These animations are a total of 10.
Example: First the ai will do a random outward hit from an array of 5 animations, then it will do the same in an array of 5 animations for the inward hit

Fixed a bug where the ai would hit itself but not deal damage. Now it will hit the player correctly

Added a feature for the shield weapon. The player can now block with the shield. It will decrease the damage taken by 50%, and take 10x of the damage amount to the stamina. If the stamina reaches 0, it will cancel the block, and the player cannot block again until the stamina reaches a min of 100.