RandomizeGear finished
Finished working on the Randomize gear function. I can now set a custom inventory for an actor, or opt to give them a random inventory (so far only weapons and armor works).

This works for the spawner as well. All spawned actors will have random items, but they will only have the possibility to have 1 item type from all the possible gear options
Example: Heavy -> 2handedaxe, all helmets, heavy and medium chestplate. The actor WILL have the weapon, that is a necessity, but it will have a random chance at all gear option, and can only have 1 of each type at most.

Randomize item function is half done. It will work similar to the chest, with the difference that certain item types, are going to be in a separate inventory (and displayed as a pouch on the mesh). When the ai dies, the pouch will drop, and act as a chest, until its empty, then it will destroy. Optionally, the player can open the inventory while its on the enemy