ItemPouch, bugfixes
Finished the itempouch. It works like the chest actor. All the items that cannot be displayed on the character, will be in this itempouch. when the enemy dies, it will drop like any other item, upon interacting with it, it will open the chest ui elements with their functions (take item from chest to player inventory, place item to chest from player inventory). If the pouch becomes empty, it will destroy itself.

Fixed bugs around enemies not dropping the items upon death, the items not being interactables etc.

Minor Ui work