Randomize gear + items finished
Finished the randomize gear function. Now it will work on items, not just gear.

Summary: I can now place an individual actor into the world, set it to custom (will only use the items given to it by hand), or random (will use the items given to it, and will have a random inventory). This works on spawner as well

All the gear has the same amount of chance to be given to the enemy based on how many items they can equip.
All items have a different chance of spawning on the enemy based on their rarity (given to it by integer for now). Some enemies could have 3 items in their pouch, some might not even have a pouch.